+++ changelog.md
@@ -2,6 +2,25 @@
 
 ## decker 0.14.0
 
+- Charts can now use the deck palette or a palette defined in YAML.
+  For this to work, disable the color plugin of Chart.js:
+  ``` yaml
+  chart:
+    defaults:
+      plugins:
+        colors:
+          enabled: false
+  ```
+  A custom palette can be defined in YAML. Colors have to be specified in HEX.
+  ``` yaml
+  chart:
+    colors:
+      - #ff0000
+      - #00ff00
+      - #00ff00
+      ...
+  ```
+
 - The colorscheme ("light" or "dark") can now be enforced in YAML. Simply add
     ``` yaml
     colorscheme: "light"